ft(register): enable users to signup 	- Enable users to register with the app 	- Enable users to login to the app 	- Add Landing and Home pages
[Finishes #173106287]
