+++ books-app/src/components/container/auth/Login.jsx
@@ -1,0 +1,69 @@
+import React, { Component } from 'react';
+import { withRouter } from 'react-router-dom';
+import { connect } from 'react-redux';
+import PropTypes from 'prop-types';
+import LoginPage from '../../presentational/auth/Login';
+import { loginUser } from '../../../actions/AuthActions';
+
+export class Login extends Component {
+  constructor() {
+    super();
+    this.state = {
+      username: '',
+      password: ''
+    };
+    this.onChange = this.onChange.bind(this);
+    this.onSubmit = this.onSubmit.bind(this);
+  }
+
+  UNSAFE_componentWillReceiveProps(nextProps) {
+    if (nextProps.errors) {
+      if (nextProps.errors.email) {
+        this.setState({ emailError: nextProps.errors.email });
+      } else if (nextProps.errors.password) {
+        this.setState({ passwordError: nextProps.errors.password });
+      } else {
+        this.setState({ notFoundUser: nextProps.errors.error });
+      }
+    } else {
+      window.localStorage.setItem('token', loginUser.nextProps.user.token);
+    }
+  }
+
+  onChange(e) {
+    this.setState({ [e.target.name]: e.target.value });
+  }
+
+  onSubmit(e) {
+    e.preventDefault();
+    const userData = {
+      Username: this.state.username,
+      Password: this.state.password
+    };
+    this.props.loginUser(userData, this.props.history);
+  }
+
+  render() {
+    return (
+      <LoginPage props={this.props} />
+    );
+  }
+}
+
+Login.propType = {
+  loginUser: PropTypes.func.isRequired,
+  auth: PropTypes.object.isRequired,
+  errors: PropTypes.object.isRequired
+};
+
+const mapStateToProps = state => {
+  return {
+    auth: state.auth,
+    errors: state.errors
+  };
+};
+
+export default connect(
+  mapStateToProps,
+  { loginUser }
+)(withRouter(Login));
